+++ app/javascript/users/store/modules/layout.js
@@ -9,7 +9,6 @@ export const state = {
   sidebarOpened: false,
   alerts: [],
   snackbar: {show: false, text: null, color: null, timeout: 3000, multi_line: true},
-  search: null,
 }
 
 export const mutations = {
@@ -32,9 +31,6 @@ export const mutations = {
     state.alerts = state.alerts.filter(function(alert) {
       return alert.id !== alertId
     })
-  },
-  SET_SEARCH(state, search) {
-    state.search = search
   },
   SET_SNACKBAR(state, data) {
     Object.assign(state.snackbar, data)
@@ -72,8 +68,5 @@ export const actions = {
   },
   setSnackbar({ commit }, data) {
     commit('SET_SNACKBAR', Object.assign(data, {show: true}))
-  },
-  setSearch({ commit }, search) {
-    commit('SET_SEARCH', search)
   },
 }